Use Django override_settings for settings fixture
This PR uses Django override_settings for the settings fixture.
This PR is mostly based on the work of PR [#141](https://github.com/pytest-dev/pytest-django/pull/324)
and more recently PR [#324](https://github.com/pytest-dev/pytest-django/pull/324).

The latter attempted to use a different mechanisim, but as discussed,
it's better to use the API provided by Django to ensure future
compatability.