Fix classmethod test
This test was incorrect in the case where the method was inherited from a
subclass.  Which ironically is precisely what the function was supposed to
test.

Additionally, there was a problem where the restored methods were bound to
the wrong class.  Added test case and fixed that also.

Fixes pytest-dev/pytest-django#597