Use pathlib2
The dependency on `pathlib` was added in #631. A backport is needed for Python <3.4.

The `pathlib` package [on PyPI](https://pypi.org/project/pathlib/) is maintenance-only and comes with a message about using `pathlib2` for an up-to-date version, still importable as `pathlib`. Also `pytest` itself uses `pathlib2` so currently I'm getting both packages installed, one on top of the other, woops!

This change in `setup.py` seems to be all that's needed to switch.