Fix DB renaming fixture for Multi-DB environment
For Django configurations using multiple databases, there's a subtle bug
that doesn't rename databases, if there's an SQLite one that triggers
the `return` statement in the existing feature.

As the return value doesn't seem to be used, we can just move to the
next iteration in the `for`-loop, and evaluate the following DB.

A test was added to consider this scenario.