Add django_debug_sql ini option
TODO:

- [ ] doc
- [ ] use a command line option instead?

I've thought about having a fixture also, but it would require to set
"force_debug_cursor" on the connections then, and it probably not
useful; typically you want to use this for a short time only - therefore
a command line option might be better suited also (but you can use `-o
django_debug_sql = 1`).  And on the other hand, it will only show up on
failures, and is therefore maybe good to set it in general.