Unblock DB in case of error for reports
This mainly works around pytest causing an INTERNALERROR for when using
`repr()` on the `pytest.fail.Exception` that is used when the DB is
blocked.

While this will be fixed via
https://github.com/pytest-dev/pytest/pull/6047 likely, it might be good
to have this anyway (also for older pytest versions).

On the other hand this might cause the DB to be accessed/changed then,
which might show different results from when the test failed due to
this?!

Fixes https://github.com/pytest-dev/pytest-django/issues/713.
Fixes https://github.com/pytest-dev/pytest-django/issues/341.