tests: Use realistic custom user model
The previous test setup inherited from AbstractUser thus MyCustomUser
still had a username field.

The problems people are having when using the admin_client fixture in
combination with a custom user model are due to the username field
not being present.

This change accounts for the more realistic scenario.

See these tickets:
https://github.com/pytest-dev/pytest-django/pull/246
https://github.com/pytest-dev/pytest-django/pull/484
https://github.com/pytest-dev/pytest-django/pull/748